add fallback skeleton
