+++ package.json
@@ -23,6 +23,7 @@
     "nprogress": "^0.2.0",
     "react": "18.2.0",
     "react-awesome-slider": "^4.1.0",
+    "react-content-loader": "^6.2.1",
     "react-dom": "18.2.0",
     "superjson": "1.9.1",
     "zod": "^3.18.0"
