+++ src/app/projects/head.tsx
@@ -1,0 +1,11 @@
+export default function Head() {
+  return (
+    <>
+      <title>Projects | FerTostado</title>
+      <meta
+        name="description"
+        content="This is a list of all the relevant projects I've worked on."
+      />
+    </>
+  );
+}
