+++ src/app/projects/page.tsx
@@ -1,8 +1,10 @@
 import { Suspense } from 'react';
 
 import { PageHeader } from '~/components/atoms/pageHeader';
+import { PageSkeleton } from '~/components/atoms/pageSkeleton';
 import { Container } from '~/components/organisms/container';
 import Projects from '~/components/containers/projects';
+
 
 
 import { getProjects } from '~/utils/projects';
@@ -17,7 +19,7 @@ export default async function ProjectsPage() {
           title="Projects"
           description="A collection of projects I've worked on."
         />
-        <Suspense fallback={'LOADING'}>
+        <Suspense fallback={<PageSkeleton />}>
           {/* @ts-expect-error Server Component */}
           <Projects />
         </Suspense>
